Game settings new UI (WIP)

+++ Assets/Scripts/Game/GameStateController.cs
@@ -125,7 +125,7 @@ namespace Game
             CleanupClientRpc();
             _roundsElapsed++;
             ResetTimer();
-            if (_roundsElapsed < _gameSettings.RoundsAmount)
+            if (_roundsElapsed < (int)_gameSettings.RoundsAmount)
             {
                 print($"Elapsed: {_roundsElapsed}, Total: {_gameSettings.RoundsAmount}");
                 LoadNextRoundClientRpc();
